requirements.txt
streamlit>=1.34.0
openai>=1.25.0
pypdf2
faiss-cpu
tiktoken
python-dotenv
pdfplumber